Backend/requirements.txt
PyMuPDF==1.23.8
spacy==3.7.2
pandas==2.1.4
numpy==1.24.3
langchain==0.1.0
langchain-community==0.0.10
faiss-cpu==1.7.4
chromadb==0.4.18
openai==1.3.7
transformers==4.36.2
torch==2.1.2
neo4j==5.15.0
beautifulsoup4==4.12.2
requests==2.31.0 